add tests for path and reuse code from existing tests
